src/requirements.txt
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5.1
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.4
django-heroku==0.3.1
djangorestframework==3.11.0
gcloud==0.17.0
googleapis-common-protos==1.51.0
heroku==0.1.4
httplib2==0.17.1
idna==2.9
isort==4.3.21
jws==0.1.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
nodejs==0.1.1
numpy==1.18.2
oauth2client==3.0.0
odict==1.7.0
optional-django==0.1.0
pandas==1.0.3
plumber==1.5
powerline-gitstatus==1.3.1
powerline-status==2.7
protobuf==3.11.3
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycryptodome==3.4.3
pylint==2.4.4
Pyrebase==3.0.27
python-dateutil==2.8.1
python-jwt==2.0.1
pytz==2019.3
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
wget==3.2
whitenoise==5.0.1
wrapt==1.11.2
zope.component==4.6.1
zope.deferredimport==4.3.1
zope.deprecation==4.4.0
zope.event==4.4
zope.hookable==5.0.1
zope.interface==5.0.1
zope.lifecycleevent==4.3
zope.proxy==4.3.5
